+++ src/main/java/com/thoughtworks/security/scpapi/service/SecurityToolService.java
@@ -1,6 +1,6 @@
 package com.thoughtworks.security.scpapi.service;
 
-import com.thoughtworks.security.scpapi.entity.SecurityTool;
+import com.thoughtworks.security.scpapi.domain.SecurityTool;
 
 import java.util.List;
 
@@ -8,7 +8,7 @@ public interface SecurityToolService {
 
     SecurityTool create(SecurityTool securityTool);
 
-    SecurityTool findById(Integer id);
+    SecurityTool findById(Long id);
 
     List<SecurityTool> findAll();
 }
